Vehicle List
This is a Single Page Application, in which the data has been fetched from an API and the details of the data present in that API are shown in UI.
In this application favorite portion has been added in which if we select a data as a favorite, the data will be shown in the Favorite option and that data will be stored in a Backend database.
We can also delete the data from the webpage.